+++ cvxpy/tests/test_examples.py
@@ -396,14 +396,13 @@ class TestExamples(BaseTest):
         ########################################
 
         # Replace the objective.
-        prob = Problem(Maximize(x + y), prob.constraints)
+        prob.objective = Maximize(x + y)
         print "optimal value", prob.solve()
 
         self.assertAlmostEqual(prob.value, 1.0)
 
         # Replace the constraint (x + y == 1).
-        constraints[0] = (x + y <= 3)
-        prob = Problem(prob.objective, constraints)
+        prob.constraints[0] = (x + y <= 3)
         print "optimal value", prob.solve()
 
         self.assertAlmostEqual(prob.value, 3.0)